# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: منصة استاذي التعليمية - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers:
    get:
      operationId: getTeachers
      tags: [teachers]
      summary: List all teachers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: subjectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Assistants
  /assistants:
    get:
      operationId: getAssistants
      tags: [assistants]
      summary: List all assistants
      parameters:
        - name: teacherId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assistants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assistant"
    post:
      operationId: createAssistant
      tags: [assistants]
      summary: Create an assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
  /assistants/{id}:
    patch:
      operationId: updateAssistant
      tags: [assistants]
      summary: Update an assistant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
    delete:
      operationId: deleteAssistant
      tags: [assistants]
      summary: Delete an assistant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Parents
...
components:
  schemas:
...
    Teacher:
      type: object
      required: [id, fullName, phone, username, bio, subjectIds, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        bio:
          type: string
        avatarUrl:
          type: ["string", "null"]
        subjectIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TeacherInput:
      type: object
      required: [fullName, phone, username, password, bio, subjectIds]
      properties:
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        subjectIds:
          type: array
          items:
            type: integer
    TeacherUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        subjectIds:
          type: array
          items:
            type: integer
        isActive:
          type: boolean

    # Assistant
    Assistant:
      type: object
      required: [id, fullName, phone, username, teacherId, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AssistantInput:
      type: object
      required: [fullName, phone, username, password, teacherId]
      properties:
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        teacherId:
          type: integer
    AssistantUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
        teacherId:
          type: integer
        isActive:
          type: boolean

    # Parent
...